version 0.4.3 become integer field

+++ fetcher/models.py
@@ -54,7 +54,7 @@ class Activity(models.Model):
 
 
 class Page(models.Model):
-    page = models.SmallIntegerField()
+    page = models.IntegerField()
     finished = models.BooleanField(default=False)
 
     def __str__(self):
